fixed stats page
